Changes to concAtTop function for regions.
Concordance for regions: When same.id=FALSE, we sort the regions by
coverage and then check to see whether a given region in the query
matches with any other regions in the query. If so, these other
regions are dropped.  We do the same procedure for the subject ranges.
Finally, we assess concordance of the 'regions'.